repository, controller, service 폴더 추가 및 schoolService 작업완료
