+++ milestone/src/main/java/com/app/milestone/repository/SchoolCustomRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface SchoolCustomRepository {
-    public List<School> findAllByDonationCount (Pageable pageable);
+//    public List<School> findAllByDonationCount (Pageable pageable);
+    public List<School> findAllByDonationCount ();
     public List<School> findAllByCreatedDate (Pageable pageable, Search search);
 }
