+++ milestone/src/main/java/com/app/milestone/repository/service/MoneyService.java
@@ -1,9 +1,9 @@
-package com.app.milestone.service;
+package com.app.milestone.repository.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ServiceService {
+public class MoneyService {
 }
