--- milestone/src/main/java/com/app/milestone/service/NoticeService.java
@@ -1,9 +1,0 @@
-package com.app.milestone.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class NoticeService {
-}
