+++ milestone/src/test/java/com/app/milestone/controller/SchoolControllerTest.java
@@ -1,0 +1,4 @@
+package com.app.milestone.controller;
+
+public class SchoolControllerTest {
+}
