+++ milestone/src/test/java/com/app/milestone/repository/money/MoneyTest.java
@@ -1,16 +1,8 @@
-package com.app.milestone.money;
+package com.app.milestone.repository.money;
 
-import com.app.milestone.domain.MoneyDTO;
-import com.app.milestone.domain.NoticeDTO;
-import com.app.milestone.entity.Money;
-import com.app.milestone.entity.Notice;
-import com.app.milestone.entity.User;
-import com.app.milestone.repository.DonationRepository;
 import com.app.milestone.repository.MoneyRepository;
-import com.app.milestone.repository.UserRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
