+++ milestone/src/test/java/com/app/milestone/repository/notice/NoticeTest.java
@@ -1,4 +1,4 @@
-package com.app.milestone.notice;
+package com.app.milestone.repository.notice;
 
 import com.app.milestone.domain.NoticeDTO;
 import com.app.milestone.entity.Notice;
