+++ milestone/src/test/java/com/app/milestone/service/SchoolServiceTest.java
@@ -1,0 +1,4 @@
+package com.app.milestone.service;
+
+public class SchoolServiceTest {
+}
